ai-news-sum-backend/requirements.txt
flask==3.0.2
playwright==1.41.2
newspaper3k==0.2.8
transformers==4.37.2
torch==2.2.0
vaderSentiment==3.3.2
python-dotenv==1.0.1
gunicorn==21.2.0
flask-cors